fix: DD-47 Code refactoring
