+++ components/Picker/TimePicker.tsx
@@ -7,7 +7,7 @@ import { nomineePlayTimeValue } from "@/states/Schedule/atom";
 const TimePicker = () => {
   const [nomineePlayTime, setNomineePlayTime] =
     useRecoilState(nomineePlayTimeValue);
-  const hours = Array.from({ length: 6 }, (_, index) => index + 1);
+  const hours = Array.from({ length: 10 }, (_, index) => index + 1);
 
   const handleDurationChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
     setNomineePlayTime(e.target.value);
@@ -18,16 +18,11 @@ const TimePicker = () => {
       <span className="w-[40px]">시간</span>
       <div className="centered-content w-[270px]">
         <SVG_down />
-        <span>09:00 -</span>
-
+        <span>00:00 -</span>
         <SVG_down />
         <span className="">24:00</span>
-
-        <label htmlFor="zz">
-          <SVG_down />
-        </label>
+        <SVG_down />
         <select
-          id="zz"
           className="appearance-none border-none"
           value={nomineePlayTime}
           onChange={handleDurationChange}
