+++ components/TestTodo/Todos.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useQuery } from "@tanstack/react-query";
-import axiosInstance from "../Axiosinstance";
+import axiosInstance from "../Axiosinstance/axiosInstance";
 
 import "dotenv/config";
