+++ containers/schedule/ScheduleIntro/index.tsx
@@ -35,7 +35,8 @@ const ScheduleIntro: React.FC<ScheduleIntroType> = ({
           </div>
         </div>
       </div>
-      <div className="calendar-scheduleIntro-layout">달력위치</div>
+      <div className="calendar-scheduleIntro-layout"></div>
+
       <Login />
     </div>
   );
