+++ package.json
@@ -13,7 +13,6 @@
     "@tanstack/react-query-devtools": "^5.14.5",
     "axios": "^1.6.4",
     "bcrypt": "^5.1.1",
-    "daisyui": "^4.4.23",
     "date-fns": "^3.1.0",
     "dotenv": "^16.3.1",
     "jsonwebtoken": "^9.0.2",
